Issue 332: Pager simplification
