+++ README.txt
@@ -4,7 +4,7 @@ Welcome to TightBlog, which started off as a fork of Apache Roller in May 2015. 
 
 Check my blog at https://web-gmazza.rhcloud.com/blog/category/Blogs+%26+Wikis for recent status updates.
 
-As of 20 January 2016, Tightblog uses 18 database tables compared to Roller's 33, and has slimmed down to 260 Java source files from the 493 in Roller.
+As of 20 February 2016, Tightblog uses 18 database tables compared to Roller's 33, and has slimmed down to 255 Java source files from the 493 in Roller.
 
 Differences between TightBlog and Roller can be seen from the closed issues page:
   https://github.com/gmazza/tightblog/issues?q=is%3Aissue+is%3Aclosed
